RC car base initialization/Added count to command
RC car base initialization was moved out of function to just before main while loop in Raspberry pi main code. A count was added in Raspberry Pi main code (as well as very minor changes to main.py for Atlas 200DK code) to modify how the robot responds to commands, some number count_num (I set it to 3) of the same hand gesture command must be detected in a row for the robot to execute the corresponding command.